feat: Update error demo
