+++ apps/insight-viewer-demo/src/containers/Error/Tabs.tsx
@@ -1,0 +1,26 @@
+import { useState } from 'react'
+import { Tabs, TabList, TabPanels, Tab, TabPanel } from '@chakra-ui/react'
+import Base from './Base'
+import Custom from './Custom'
+
+export default function CustomError(): JSX.Element {
+  const [active, setActive] = useState(0)
+
+  function handleChange(index: number): void {
+    setActive(index)
+  }
+
+  return (
+    <Tabs isLazy onChange={handleChange}>
+      <TabList>
+        <Tab>Default Error</Tab>
+        <Tab>Custom Error</Tab>
+      </TabList>
+
+      <TabPanels>
+        <TabPanel>{active === 0 && <Base />}</TabPanel>
+        <TabPanel>{active === 1 && <Custom />}</TabPanel>
+      </TabPanels>
+    </Tabs>
+  )
+}
